+++ AppLovin/src/main/java/com/mopub/mobileads/AppLovinInterstitial.java
@@ -16,7 +16,6 @@ import com.applovin.sdk.AppLovinAdDisplayListener;
 import com.applovin.sdk.AppLovinAdLoadListener;
 import com.applovin.sdk.AppLovinAdSize;
 import com.applovin.sdk.AppLovinAdVideoPlaybackListener;
-import com.applovin.sdk.AppLovinErrorCodes;
 import com.applovin.sdk.AppLovinMediationProvider;
 import com.applovin.sdk.AppLovinPrivacySettings;
 import com.applovin.sdk.AppLovinSdk;
@@ -59,9 +58,10 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
     private static final Map<String, Queue<AppLovinAd>> GLOBAL_INTERSTITIAL_ADS = new HashMap<String, Queue<AppLovinAd>>();
     private static final Object GLOBAL_INTERSTITIAL_ADS_LOCK = new Object();
 
-    private String zoneId; // The zone identifier this instance of the custom event is loading for
     private boolean isTokenEvent;
     private AppLovinAd tokenAd;
+
+    private static String mZoneId; // The zone identifier this instance of the custom event is loading for
 
     @NonNull
     private AppLovinAdapterConfiguration mAppLovinAdapterConfiguration;
@@ -76,6 +76,13 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
 
     @Override
     public void loadInterstitial(final Context context, final CustomEventInterstitialListener listener, final Map<String, Object> localExtras, final Map<String, String> serverExtras) {
+        if (serverExtras == null || serverExtras.isEmpty()) {
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "No serverExtras provided");
+            if (listener != null) {
+                listener.onInterstitialFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+            }
+            return;
+        }
 
         // Pass the user consent from the MoPub SDK to AppLovin as per GDPR
         boolean canCollectPersonalInfo = MoPub.canCollectPersonalInformation();
@@ -83,10 +90,10 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
 
         // SDK versions BELOW 7.2.0 require a instance of an Activity to be passed in as the context
         if (AppLovinSdk.VERSION_CODE < 720 && !(context instanceof Activity)) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Unable to request AppLovin interstitial. Invalid context " +
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Unable to request AppLovin interstitial. Invalid context " +
                     "provided.");
 
-            MoPubLog.log(LOAD_FAILED, ADAPTER_NAME,
+            MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
                     MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(),
                     MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
             if (listener != null) {
@@ -102,9 +109,9 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
         sdk = retrieveSdk(context);
 
         if (sdk == null) {
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "AppLovinSdk instance is null likely because " +
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "AppLovinSdk instance is null likely because " +
                     "no AppLovin SDK key is available. Failing ad request.");
-            MoPubLog.log(LOAD_FAILED, ADAPTER_NAME,
+            MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
                     MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(),
                     MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
 
@@ -121,7 +128,7 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
         final String adMarkup = serverExtras.get(DataKeys.ADM_KEY);
         final boolean hasAdMarkup = !TextUtils.isEmpty(adMarkup);
 
-        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Requesting AppLovin interstitial with serverExtras: " +
+        MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Requesting AppLovin interstitial with serverExtras: " +
                 serverExtras + ", localExtras: " + localExtras + " and has adMarkup: " + hasAdMarkup);
 
         mAppLovinAdapterConfiguration.setCachedInitializationParameters(context, serverExtras);
@@ -131,25 +138,25 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
 
             // Use token API
             sdk.getAdService().loadNextAdForAdToken(adMarkup, this);
-            MoPubLog.log(zoneId, LOAD_ATTEMPTED, ADAPTER_NAME);
+            MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
         } else {
             final String serverExtrasZoneId = serverExtras.get(ZONE_ID_SERVER_EXTRAS_KEY);
-            zoneId = !TextUtils.isEmpty(serverExtrasZoneId) ? serverExtrasZoneId : DEFAULT_ZONE;
+            mZoneId = !TextUtils.isEmpty(serverExtrasZoneId) ? serverExtrasZoneId : DEFAULT_ZONE;
 
             // Check if we already have a preloaded ad for the given zone
-            final AppLovinAd preloadedAd = dequeueAd(zoneId);
+            final AppLovinAd preloadedAd = dequeueAd(mZoneId);
             if (preloadedAd != null) {
-                MoPubLog.log(CUSTOM, ADAPTER_NAME, "Found preloaded ad for zone: {" + zoneId + "}");
+                MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Found preloaded ad for zone: {" + mZoneId + "}");
                 adReceived(preloadedAd);
             }
             // No ad currently preloaded
             else {
-                if (!TextUtils.isEmpty(zoneId)) {
-                    sdk.getAdService().loadNextAdForZoneId(zoneId, this);
-                    MoPubLog.log(zoneId, LOAD_ATTEMPTED, ADAPTER_NAME);
+                if (!TextUtils.isEmpty(mZoneId)) {
+                    sdk.getAdService().loadNextAdForZoneId(mZoneId, this);
+                    MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
                 } else {
                     sdk.getAdService().loadNextAd(AppLovinAdSize.INTERSTITIAL, this);
-                    MoPubLog.log(zoneId, LOAD_ATTEMPTED, ADAPTER_NAME);
+                    MoPubLog.log(getAdNetworkId(), LOAD_ATTEMPTED, ADAPTER_NAME);
                 }
             }
         }
@@ -159,12 +166,12 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
     public void showInterstitial() {
         final AppLovinAd preloadedAd;
 
-        MoPubLog.log(SHOW_ATTEMPTED, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), SHOW_ATTEMPTED, ADAPTER_NAME);
 
         if (isTokenEvent && tokenAd != null) {
             preloadedAd = tokenAd;
         } else {
-            preloadedAd = dequeueAd(zoneId);
+            preloadedAd = dequeueAd(mZoneId);
         }
 
         if (preloadedAd != null) {
@@ -175,14 +182,14 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
             interstitialAd.setAdVideoPlaybackListener(this);
             interstitialAd.showAndRender(preloadedAd);
         } else {
-            MoPubLog.log(SHOW_FAILED, ADAPTER_NAME, MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR.getIntCode(),
-                    MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+            MoPubLog.log(getAdNetworkId(), SHOW_FAILED, ADAPTER_NAME, MoPubErrorCode.NETWORK_NO_FILL.getIntCode(),
+                    MoPubErrorCode.NETWORK_NO_FILL);
 
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Failed to show an AppLovin interstitial before one was " +
+            MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Failed to show an AppLovin interstitial before one was " +
                     "loaded");
 
             if (listener != null) {
-                listener.onInterstitialFailed(MoPubErrorCode.ADAPTER_CONFIGURATION_ERROR);
+                listener.onInterstitialFailed(MoPubErrorCode.NETWORK_NO_FILL);
             }
         }
     }
@@ -201,20 +208,20 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
         if (isTokenEvent) {
             tokenAd = ad;
         } else {
-            enqueueAd(ad, zoneId);
+            enqueueAd(ad, mZoneId);
         }
 
         runOnUiThread(new Runnable() {
             @Override
             public void run() {
                 try {
-                    MoPubLog.log(LOAD_SUCCESS, ADAPTER_NAME);
+                    MoPubLog.log(getAdNetworkId(), LOAD_SUCCESS, ADAPTER_NAME);
 
                     if (listener != null) {
                         listener.onInterstitialLoaded();
                     }
                 } catch (Throwable th) {
-                    MoPubLog.log(CUSTOM_WITH_THROWABLE, "Unable to notify listener of " +
+                    MoPubLog.log(getAdNetworkId(), CUSTOM_WITH_THROWABLE, "Unable to notify listener of " +
                             "successful ad load", th);
                 }
             }
@@ -228,15 +235,15 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
             @Override
             public void run() {
                 try {
-                    MoPubLog.log(LOAD_FAILED, ADAPTER_NAME,
-                            toMoPubErrorCode(errorCode).getIntCode(),
-                            toMoPubErrorCode(errorCode));
+                    MoPubLog.log(getAdNetworkId(), LOAD_FAILED, ADAPTER_NAME,
+                            AppLovinAdapterConfiguration.getMoPubErrorCode(errorCode).getIntCode(),
+                            AppLovinAdapterConfiguration.getMoPubErrorCode(errorCode));
 
                     if (listener != null) {
-                        listener.onInterstitialFailed(toMoPubErrorCode(errorCode));
+                        listener.onInterstitialFailed(AppLovinAdapterConfiguration.getMoPubErrorCode(errorCode));
                     }
                 } catch (Throwable th) {
-                    MoPubLog.log(CUSTOM_WITH_THROWABLE, "Unable to notify listener of failure" +
+                    MoPubLog.log(getAdNetworkId(), CUSTOM_WITH_THROWABLE, "Unable to notify listener of failure" +
                             " to receive ad", th);
                 }
             }
@@ -249,7 +256,7 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
 
     @Override
     public void adDisplayed(final AppLovinAd appLovinAd) {
-        MoPubLog.log(SHOW_SUCCESS, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), SHOW_SUCCESS, ADAPTER_NAME);
 
         if (listener != null) {
             listener.onInterstitialShown();
@@ -269,7 +276,7 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
 
     @Override
     public void adClicked(final AppLovinAd appLovinAd) {
-        MoPubLog.log(CLICKED, ADAPTER_NAME);
+        MoPubLog.log(getAdNetworkId(), CLICKED, ADAPTER_NAME);
 
         if (listener != null) {
             listener.onInterstitialClicked();
@@ -282,13 +289,17 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
 
     @Override
     public void videoPlaybackBegan(final AppLovinAd ad) {
-        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Interstitial video playback began");
+        MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Interstitial video playback began");
     }
 
     @Override
     public void videoPlaybackEnded(final AppLovinAd ad, final double percentViewed, final boolean fullyWatched) {
-        MoPubLog.log(CUSTOM, ADAPTER_NAME, "Interstitial video playback ended at playback percent: ",
+        MoPubLog.log(getAdNetworkId(), CUSTOM, ADAPTER_NAME, "Interstitial video playback ended at playback percent: ",
                 percentViewed);
+    }
+
+    private static String getAdNetworkId() {
+        return mZoneId;
     }
 
     //
@@ -315,20 +326,6 @@ public class AppLovinInterstitial extends CustomEventInterstitial implements App
                 GLOBAL_INTERSTITIAL_ADS.put(zoneId, preloadedAds);
             }
             preloadedAds.offer(ad);
-        }
-    }
-
-    private static MoPubErrorCode toMoPubErrorCode(final int applovinErrorCode) {
-        if (applovinErrorCode == AppLovinErrorCodes.NO_FILL) {
-            return MoPubErrorCode.NETWORK_NO_FILL;
-        } else if (applovinErrorCode == AppLovinErrorCodes.UNSPECIFIED_ERROR) {
-            return MoPubErrorCode.UNSPECIFIED;
-        } else if (applovinErrorCode == AppLovinErrorCodes.NO_NETWORK) {
-            return MoPubErrorCode.NO_CONNECTION;
-        } else if (applovinErrorCode == AppLovinErrorCodes.FETCH_AD_TIMEOUT) {
-            return MoPubErrorCode.NETWORK_TIMEOUT;
-        } else {
-            return MoPubErrorCode.UNSPECIFIED;
         }
     }